backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pypdf
python-multipart
supabase
psycopg2-binary
pyspark
numpy
httpx
beautifulsoup4
lxml

